+++ src/object_avoidance/test/data/obstacle_front.bagy
